fix: prevented editing entity when it doesn't exist in DB, added validation group constraints

+++ src/main/java/io/github/batetolast1/springcms/service/impl/DefaultArticleService.java
@@ -9,7 +9,6 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -25,7 +24,7 @@ public class DefaultArticleService implements ArticleService {
     @Override
     public List<ArticleDto> getLastArticles() {
         return articleDao
-                .findFirst5ByOrderByCreatedOnDesc()
+                .findFirst5ByDraftFalseByOrderByCreatedOnDesc()
                 .stream()
                 .map(a -> modelMapper.map(a, ArticleDto.class))
                 .collect(Collectors.toList());
@@ -43,25 +42,39 @@ public class DefaultArticleService implements ArticleService {
 
     @Override
     public void delete(Long id) {
-        articleDao.delete(id);
+        if (existsAndIsArticle(id)) {
+            articleDao.delete(id);
+        }
     }
 
     @Override
     public void save(ArticleDto articleDto) {
         Article article = modelMapper.map(articleDto, Article.class);
-        article.setCreatedOn(LocalDateTime.now());
+        article.setDraft(false);
         articleDao.save(article);
     }
 
     @Override
     public ArticleDto getById(Long id) {
-        Article article = articleDao.findById(id);
-        return modelMapper.map(article, ArticleDto.class);
+        if (existsAndIsArticle(id)) {
+            Article article = articleDao.findById(id);
+            return modelMapper.map(article, ArticleDto.class);
+        }
+
+        return null;
     }
 
     @Override
     public void edit(ArticleDto articleDto) {
         Article article = modelMapper.map(articleDto, Article.class);
-        articleDao.update(article);
+
+        if (existsAndIsArticle(article.getId())) {
+            articleDao.update(article);
+        }
+    }
+
+    private boolean existsAndIsArticle(Long id) {
+        Article article = articleDao.findById(id);
+        return article != null && !article.getDraft();
     }
 }
